+++ Gemfile
@@ -1,6 +1,8 @@
 source 'https://rubygems.org'
 
 #権限管理
+gem 'banken'
+gem 'cancancan'
 gem "pundit"
 #日本語化
 gem 'devise-i18n'
