+++ app/loyalties/users_loyalty.rb
@@ -1,0 +1,8 @@
+class UsersLoyalty < ApplicationLoyalty
+  
+ def update?
+    user.role == 'admin' || user.id == record.id
+  end
+
+
+end
