+++ app/policies/user_policy.rb
@@ -1,0 +1,8 @@
+#class UserPolicy < ApplicationPolicy
+# 
+#  
+#  def update?
+#    role.admin? 
+#  end
+#  
+#end
