+++ app/views/devise/sessions/new.html.erb
@@ -17,10 +17,10 @@
                       <%= f.password_field :password, class: 'form-control', autocomplete: "off" %>
                     </div>
 
-                    <div class="form-group">
-                        <%= f.label :role %><br />
-                        <%= f.select :role ,[["管理者", "admin"], ["ユーザ", "user"]] , class: 'form-control', autofocus: true %>
-                      </div>
+                    <%#*<div class="form-group">%>
+                        <%#= f.label :role %><br />
+                        <%#= f.select :role ,[["管理者", "admin"], ["ユーザ", "user"]] , class: 'form-control', autofocus: true %>
+                      <%#*</div>%>
 
                       <% if devise_mapping.rememberable? -%>
                         <div class="form-group" >
